Updates so that vagrant can work on slower hardware
The following updates ensure that the vagrant tests
can work on slower hardware than what I was using before.

Updates the amount of memory per VM (on worse hardware,
things start thrashing in a way that leads to build failures)

Increase the sleep time to give it more time to set up
networking so that the floating ip allocation will actually
work.